changed uint <-> double casting

+++ mdl_people_tracker/include/Detections.h
@@ -56,11 +56,11 @@ public:
     double getHeight( int frame,  int detec);
 //    void getDetection(int frame, int detec, Vector<double>& det);
     int getCategory(int frame, int detec);
-    int getHeaderSeq(int frame, int detec){
-    	return detC(frame)(detec)(24);
+    uint32_t getHeaderSeq(int frame, int detec){
+    	return static_cast<unsigned int>(detC(frame)(detec)(24));
     }
-    int getIndex(int frame, int detec){
-    	return detC(frame)(detec)(1);
+    uint32_t getIndex(int frame, int detec){
+    	return static_cast<unsigned int>(detC(frame)(detec)(1));
     }
 //    int getDetNumber(int frame, int detec);
 //    Vector<Vector<double> > get3Dpoints(int frame, int detec);
